+++ lib/src/server.dart
@@ -181,6 +181,27 @@ abstract class StreamServer {
    */
   ConnectErrorCallback get defaultErrorCallback;
 
+  /** Maps the given URI to the given handler.
+   *
+   * * [uri] - a regular exception used to match the request URI.
+   * * [handler] - the handler for handling the request, or another URI that this request
+   * will be forwarded to. If the value is a URI and the key has named groups, the URI can
+   * refer to the group with the $ expression.
+   * For example: `'/dead-link(info:.*)': '/new-link$info'`.
+   * if [handler] is null, it means removal.
+   * * [preceding] - whether to make the mapping preceding any previous mappings.
+   * In other words, if true, this mapping will be interpreted first.
+   */
+  void map(String uri, handler, {preceding: false});
+  /** Maps the given URI to the given filter.
+   *
+   * * [uri]: a regular exception used to match the request URI.
+   * * [filter]: the filter. If null, it means removal.
+   * * [preceding] - whether to make the mapping preceding any previous mappings.
+   * In other words, if true, this mapping will be interpreted first.
+   */
+  void filter(String uri, RequestFilter filter, {preceding: false});
+
   /** The logger for logging information.
    * The default level is `INFO`.
    */